move code into more disecrete files
